widgets from view moved to widgets
